IntegerField

+++ mainpage/models.py
@@ -40,7 +40,7 @@ class Moscow_zone(models.Model):
         return self.name
 class Cadastra_val_Value(models.Model): #Средняя кадастровая стоимость
     district = models.ForeignKey(Moscow_zone, on_delete=models.CASCADE) #район Москвы
-    cost = models.DecimalField(max_length=10000)
+    cost = models.IntegerField(max_length=10000)
 
 class Order_Varibles(models.Model): #калькулятор
     industry_type = models.ForeignKey(Branch, on_delete=models.CASCADE) #отрасль
